extracting animation-logic from the PopupAdvice-component to a customized hook
