+++ src/components/common/popup-advice/PopupAdvice.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useRef } from 'react';
-import { Animated } from 'react-native';
+import React from 'react';
 
 import * as Styles from './PopupAdvice.styles';
+import usePopupAdvice from './usePopupAdvice';
 
 export const DEFAULT_ANIMATION_DURATION = 400;
 export const HIDE_POPUP_DELAY = 3000;
@@ -12,27 +12,7 @@ type PopupAdviceProps = {
 };
 
 const PopupAdvice = ({ onFinishToShow = () => {}, text }: PopupAdviceProps) => {
-  const wrapperOpacity = useRef(new Animated.Value(0)).current;
-
-  useEffect(() => {
-    const onHidePopup = () => {
-      Animated.timing(wrapperOpacity, {
-        duration: DEFAULT_ANIMATION_DURATION,
-        useNativeDriver: true,
-        toValue: 0,
-      }).start(onFinishToShow);
-    };
-
-    Animated.timing(wrapperOpacity, {
-      duration: DEFAULT_ANIMATION_DURATION,
-      useNativeDriver: true,
-      toValue: 1,
-    }).start(() => {
-      setTimeout(() => {
-        onHidePopup();
-      }, HIDE_POPUP_DELAY);
-    });
-  }, []);
+  const { wrapperOpacity } = usePopupAdvice({ onFinishToShow });
 
   return (
     <Styles.Wrapper
